+++ src/lapack/syev/d2f.sh
@@ -1,0 +1,4 @@
+cat dense_double_syev.cpp \
+    | sed -e 's/double/float/g' \
+    | sed -e 's/dsyev/ssyev/g' \
+    > dense_float_syev.cpp